+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Сhangelog
 
+- Switched HitBTC v2 fetchBalance to trading account by default, [#210](https://github.com/kroitor/ccxt/pull/210)
 - Added an internal rate-limiting REST poller to JS (experimental)
 - Fixed Cryptopia fetchTicker volume [#206](https://github.com/kroitor/ccxt/issues/206)
 - Fixed Bittrex await cancelOrder [#205](https://github.com/kroitor/ccxt/issues/205)